+++ src/tests/EntityGraphQL.Tests/TestDataContext.cs
@@ -12,11 +12,17 @@ namespace EntityGraphQL.Tests
     public class TestDataContext
     {
         public int TotalPeople => People.Count;
+        [Obsolete("This is obsolete, use Projects instead")]
+        public IEnumerable<ProjectOld> ProjectsOld { get; set; }
         public IEnumerable<Project> Projects { get; set; }
         public IEnumerable<Task> Tasks { get; set; } = new List<Task>();
         public List<Location> Locations { get; set; } = new List<Location>();
         public virtual List<Person> People { get; set; } = new List<Person>();
         public IEnumerable<User> Users { get; set; } = new List<User>();
+    }
+
+    public class ProjectOld
+    {
     }
 
     public enum Gender